pyproject.toml
[project]
name = "worldsignal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi>=0.128.8",
    "feedparser>=6.0.12",
    "httpx>=0.28.1",
    "langchain-text-splitters>=1.1.0",
    "llama-cpp-python>=0.3.16",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.2",
    "sentence-transformers>=5.2.2",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
]
